requirements.txt
cryptography==43.0.1
pyjwt==2.8.0
argon2-cffi==23.1.0